feat: add react-router to app
