+++ package.json
@@ -14,6 +14,7 @@
     "@ya.praktikum/react-developer-burger-ui-components": "^1.13.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.1",
     "react-scripts": "^5.0.0",
     "typescript": "^4.5.4",
     "uuid": "^8.3.2",
