+++ src/components/app/app.module.css
@@ -7,7 +7,6 @@
 
 .content {
   display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
+  flex-flow: row nowrap;
   justify-content: center;
 }
