+++ src/pages/forgot-password.jsx
@@ -1,0 +1,10 @@
+const ForgotPasswordPage = () => {
+
+  return (
+    <>
+
+    </>
+  );
+}
+
+export default ForgotPasswordPage;
