+++ src/pages/index.js
@@ -1,0 +1,8 @@
+import HomePage from './home.jsx';
+import LoginPage from './login.jsx';
+import RegisterPage from './register.jsx';
+import ForgotPasswordPage from './forgot-password.jsx';
+import ResetPasswordPage from './reset-password.jsx';
+import ProfilePage from './profile.jsx';
+
+export { HomePage, LoginPage, RegisterPage, ForgotPasswordPage, ResetPasswordPage, ProfilePage };
