+++ src/pages/login.jsx
@@ -1,0 +1,10 @@
+const LoginPage = () => {
+
+  return (
+    <>
+
+    </>
+  );
+}
+
+export default LoginPage;
