+++ src/pages/profile.jsx
@@ -1,0 +1,10 @@
+const ProfilePage = () => {
+
+  return (
+    <>
+
+    </>
+  );
+}
+
+export default ProfilePage;
