+++ src/pages/register.jsx
@@ -1,0 +1,10 @@
+const RegisterPage = () => {
+
+  return (
+    <>
+
+    </>
+  );
+}
+
+export default RegisterPage;
