+++ src/pages/reset-password.jsx
@@ -1,0 +1,10 @@
+const ResetPasswordPage = () => {
+
+  return (
+    <>
+
+    </>
+  );
+}
+
+export default ResetPasswordPage;
